[WIP] Support tensors as Dict keys
This prepares the PR where we implement sets in terms of dicts.

While implementing this I had to make hashable keys somewhat consistent
all throughout. I did that by having an auxiliary class (similar to the
one that was inside SetVariable). This variable is opaque on purpose, so
that it fails hard if it unadvertedly leaks back into user code.

We also found and fixed a number of latent bugs, like where dict with
builtin types was going through to UserDefinedObjectVariable, even
though we had a test that was testing that this worked (the test was not
testing whether we were tracing anything or not).

ghstack-source-id: fe83ee33feaf75b2677f361f7ea140721742ddfa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/111196